First commit for version 2.0 Beta
First commit for version 2.0 Beta. Not final, but calculation should work as intended. ToDo: Saving data, so its diffrent from other calculator, also probably the same with the link(this is more a nice to have). Also took out, that the boxes with errors get red in mobile, because it's not working with the rows in Semester with a two diggit roman number. Have to fix it, when I'm done with the more important stuff.